Move area_test.go into geom_test.go
